Moved showTokenDialog to RoomFXController because it's a specific behaviour
